leave session and join session
