+++ frontend/src/app/(with-websocket)/rooms/[roomCode]/[sessionId]/host/page.tsx
@@ -1,0 +1,10 @@
+import { HostWaitingRoom } from "@/components/pages/rooms/host/HostWaitingRoom";
+import { FullHeightCardLayout } from "@/components/ui/FullHeightCardLayout";
+
+export default function RoomsHostPage() {
+  return (
+    <FullHeightCardLayout>
+      <HostWaitingRoom />
+    </FullHeightCardLayout>
+  );
+}
