+++ frontend/src/app/(with-websocket)/rooms/[roomCode]/page.tsx
